requirements.txt
asgiref==3.8.1
Django==5.2
django-unfold==0.55.3
djangorestframework==3.16.0
psycopg2==2.9.10
python-nmap==0.7.1
sqlparse==0.5.3
